platform/pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name        = "vs-agent-platform"
version     = "0.1.0"
description = "Vidya Sankalp Agentic Platform — FastAPI gateway, auth, rate limiting, prompt versioning"
readme      = "README.md"
requires-python = ">=3.11"

dependencies = [
    # Core foundation
    "vs-agent-core>=0.1.0",

    # API framework
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",

    # Auth
    "PyJWT[cryptography]>=2.8",
    "httpx>=0.27",

    # AWS
    "boto3>=1.34",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-html>=4.0",
    "httpx>=0.27",   # for TestClient
]

[tool.setuptools.packages.find]
where   = ["."]
include = ["platform*"]

[project.scripts]
vs-platform = "platform.main:app"

[project.urls]
Homepage   = "https://vidyasankalp.com"
Repository = "https://github.com/vidyasankalp/vs-agentic-platform"
